Improved identification of base types for magic items. The base type name for Rings, amulets, belts & potions is now correctly determined.
Fixed the order of the tests for determining item category, so items
like Ring Coif are now correctly detected as items rather than rings.

Next job is to store the prefix & suffix effects in the item object, and
then use them to infer what the item's mods are. (and the number of
them).